+++ modules/querier/config.go
@@ -11,22 +11,31 @@ import (
 
 // Config for a querier.
 type Config struct {
+	Search SearchConfig `yaml:"search"`
+
 	TraceLookupQueryTimeout time.Duration `yaml:"query_timeout"`
-	SearchQueryTimeout      time.Duration `yaml:"search_query_timeout"`
-	SearchExternalEndpoints []string      `yaml:"search_external_endpoints"`
-	SearchPreferSelf        int           `yaml:"search_prefer_self"`
 	ExtraQueryDelay         time.Duration `yaml:"extra_query_delay,omitempty"`
 	MaxConcurrentQueries    int           `yaml:"max_concurrent_queries"`
 	Worker                  worker.Config `yaml:"frontend_worker"`
 }
 
+type SearchConfig struct {
+	QueryTimeout      time.Duration `yaml:"query_timeout"`
+	PreferSelf        int           `yaml:"prefer_self"`
+	ExternalEndpoints []string      `yaml:"external_endpoints"`
+	HedgeRequestsAt   time.Duration `yaml:"external_hedge_requests_at"`
+	HedgeRequestsUpTo int           `yaml:"external_hedge_requests_up_to"`
+}
+
 // RegisterFlagsAndApplyDefaults register flags.
 func (cfg *Config) RegisterFlagsAndApplyDefaults(prefix string, f *flag.FlagSet) {
 	cfg.TraceLookupQueryTimeout = 10 * time.Second
-	cfg.SearchQueryTimeout = 30 * time.Second
 	cfg.ExtraQueryDelay = 0
 	cfg.MaxConcurrentQueries = 5
-	cfg.SearchPreferSelf = 2
+	cfg.Search.PreferSelf = 2
+	cfg.Search.HedgeRequestsAt = 4 * time.Second
+	cfg.Search.HedgeRequestsUpTo = 3
+	cfg.Search.QueryTimeout = 30 * time.Second
 	cfg.Worker = worker.Config{
 		MatchMaxConcurrency:   true,
 		MaxConcurrentRequests: cfg.MaxConcurrentQueries,